sepolicy: Update sepolicy for qseecom rpmb component
UFS is another primary storage device that is supported and has a RPMB
partition that is used by the rpmb listener in qseecom. This change
updates the selinux policy to allow access to scsi generic devices (as
current UFS devices work as scsi devices) for the rpmb component to
function as expected.

CRs-Fixed: 781763
Change-Id: I74b7d1739720b9e0d4feff04962d9dd3e16e9c20